* Fixed spurious history objects being created on document upload  * Modified DataObj to clear changed fields after commit  * Only commit in files_modified if files were actually modified  * Some background:
When a document was being uploaded two instances of the eprint object were
being created - one in the form update and another as an on-demand parent of the
new document. Both of these objects were then being committed, resulting in
two history objects for the same revision.

This may not have been an issue before because eprint objects were always being
created from the database by get_eprint(). Objects are now "cached" as the
parent of the sub-object, so we need to be more careful.

This commit reduces the number of spurious commits() and revisions to the
correct number for a document upload (i.e. one).


git-svn-id: https://svn.eprints.org/eprints/trunk/system@4313 9491667e-5006-0410-a446-efbe8990b998