Allow any-of simple search over subjects
Fixes #15

InSubject mints a new prefix in join and logic, but SubQuery (as inherited by
OrSubQuery) only uses the join from the first sub-op. Hence the perfix used in
logic for any other ORed InSubject sub-ops would not match the prefix used in
the join.

Fixed by always setting a prefix in the call to join() and logic() - empty
string so we don't make the SQL any more difficult to read than it already is.