Limit the number of items on a browse view page
Fixes #25.

Update::Views calls get_record() for the complete result set on a page. This
can break memory/processing limits when generating views on large sites.

As a crude fix we'll apply a limit to the maximum number of items on a single page. It's not clever to send huge HTML files to users anyway.