Force commit in lift_embargoes
Fixes #21.

Most of the time lift_embargoes will cause an eprint commit (/revision) due to
fileinfo changing. However if there is no icon on the document the eprint will
not be committed, causing any triggered behaviour (e.g. full_text_status) to
not occur.

This change calls a force-commit on the document (which will force-commit the
eprint in turn).