Added DataSet::owned_by() to get owned items
Added a generic concept for getting owned items from arbitrary datasets.
"owned" is a record that a user has created/controls. Returns a SearchExp to
allow consuming code to manipulate the search before executing.

Adds a DataSet property 'owned' that is a function:
$f($dataset, $user, $filters, %opts)

Where $f should manipulate $filters to add the search criteria to constrain the
matches to items owned by $user.